[mypyc] Refactor: extract function related genops to new modules
Make minimal changes.
