+++ mypyc/genopsutil.py
@@ -2,8 +2,11 @@ from typing import Dict, Any, Union, Optional
 
 from mypy.nodes import (
     ClassDef, FuncDef, Decorator, OverloadedFuncDef, StrExpr, CallExpr, RefExpr, Expression,
-    IntExpr, FloatExpr, ARG_NAMED, ARG_NAMED_OPT, ARG_POS, ARG_OPT
+    IntExpr, FloatExpr, Var, TupleExpr, UnaryExpr, BytesExpr, ARG_NAMED, ARG_NAMED_OPT, ARG_POS,
+    ARG_OPT, GDEF
 )
+from mypyc.ops import Environment, ClassIR, RInstance, AssignmentTargetRegister
+from mypyc.common import SELF_NAME
 
 
 def is_trait_decorator(d: Expression) -> bool:
@@ -102,3 +105,28 @@ def concrete_arg_kind(kind: int) -> int:
         return ARG_NAMED
     else:
         return kind
+
+
+def is_constant(e: Expression) -> bool:
+    """Check whether we allow an expression to appear as a default value.
+
+    We don't currently properly support storing the evaluated
+    values for default arguments and default attribute values, so
+    we restrict what expressions we allow.  We allow literals of
+    primitives types, None, and references to Final global
+    variables.
+    """
+    return (isinstance(e, (StrExpr, BytesExpr, IntExpr, FloatExpr))
+            or (isinstance(e, UnaryExpr) and e.op == '-'
+                and isinstance(e.expr, (IntExpr, FloatExpr)))
+            or (isinstance(e, TupleExpr)
+                and all(is_constant(e) for e in e.items))
+            or (isinstance(e, RefExpr) and e.kind == GDEF
+                and (e.fullname in ('builtins.True', 'builtins.False', 'builtins.None')
+                     or (isinstance(e.node, Var) and e.node.is_final))))
+
+
+def add_self_to_env(environment: Environment, cls: ClassIR) -> AssignmentTargetRegister:
+    return environment.add_local_reg(
+        Var(SELF_NAME), RInstance(cls), is_arg=True
+    )
